+++ src/aks-preview/HISTORY.rst
@@ -14,7 +14,7 @@ Pending
 
 0.5.126
 +++++++
-* Add `--nrg-restriction <restriction level>` option for chosing the node resource group restriction level in `aks create` and `aks update`
+* Add `--nrg-lockdown-restriction-level <restriction level>` option for chosing the node resource group restriction level in `aks create` and `aks update`
 
 0.5.125
 +++++++