+++ src/aks-preview/HISTORY.rst
@@ -13,6 +13,10 @@ Pending
 +++++++
 * Raise InvalidArgumentValueError for azure cni + pod_cidr without overlay.
 
+0.5.126
++++++++
+* Add `--nrg-lockdown-restriction-level <restriction level>` option for chosing the node resource group restriction level in `aks create` and `aks update`
+
 0.5.125
 +++++++
 * Update the minimum required cli core version to `2.44.0`.